warningの対応

+++ moqt-core/src/modules/message_handler.rs
@@ -110,7 +110,7 @@ pub async fn message_handler(
     }
 
     // ciでエラーが出るかを確認
-    let payload_length = read_cur.remaining() as usize;
+    let payload_length = read_cur.remaining();
 
     // 正しく読めたのでその分bufferを進める
     read_buf.advance(read_cur.position() as usize);
